Attempt another solver if the first does not find a solution.
Mistral seems to be the only solver capable of finding multiple
solutions but, sadly, sometimes it wrongly concludes a problem is
unsatisfiable. (I.e., when given the solution it agrees it is valid.)

Add fallback to other solvers, namely MiniSat which cannot find multiple
solutions but does seem to try harder than Mistral.