Improve dimension constraints for two ambiguous dimensions.
If two dimension will have duplicates but the number of duplicates is
not known then the best conclusion we can draw is that the total number
of "True" cells is equal to the max dimension size.

For example, if there are 7 names then it is impossible for ALL cells in
a size-2 dimension * size-4 dimension to be true because that would be
8 true values and 8 > 7.

Updated the test which should have caught this issue.