Add ngram support to trie module.
Given a list of unigrams, bigrams, trigrams, etc. produce a graph of
Tries.

A path from root through unigrams gives a weight equal to the
probability for that unigram. A path through bigrams will give the
probability that specific bigram would be chosen relative to the total
count for the unigram prefix.

Straying away from given ngrams will produce paths with greatly reduced
probabilities. An invented ngram will have probability equal to the min
probability for all of its prefixes.