Add data structure for enumerating anagram options.
A bitfield is used to manage available options. This enables set
membership and subtraction to be performed with binary operations.

If the anagram is made from individual characters the lowest order bit
is used each time. Otherwise, accessing a child by prefix will trigger
O(n!) exploration for possible combinations to cover the prefix. When
the solution is unique compaction occurs at that point: prefix is
truncated to "" and available items drops. See tests for examples.